Added logic for syncing packing lists Added a BaseSalesforce Class Added a generic pull and push fields sync utility method.

+++ src/ke/co/blueconsulting/sianroses/data/sync/CustomerContacts.java
@@ -79,7 +79,7 @@ public class CustomerContacts {
 		
 		try {
 			
-			ArrayList<Long> customerIds = new ArrayList<>();
+			ArrayList<Integer> customerIds = new ArrayList<>();
 			
 			for (CustomerContact customerContact : customerContacts) {
 				customerIds.add(customerContact.getAutoId());
